Fix JPorta errors
Refactor fresnelReflectance and add clampp function

Introduced a new template function `clampp` to replace the previous `clamp` usage in the `fresnelReflectance` method of the `Material` struct. Added temporary variables for improved readability in the calculations of `rParallel2` and `rPerp2`. Modified the return value of `fresnelReflectance` to return `outRad * 0.5f`, impacting the final output.